pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "erp-docs-mirror"
version = "0.1.0"
description = "Mirror ERPNext and Frappe docs into local Markdown wiki format"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "html2text>=2024.2.26",
  "httpx>=0.27.0",
  "PyYAML>=6.0.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
